% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabraux.R
\name{check_times}
\alias{check_times}
\title{check_times}
\usage{
check_times(times, steps = 8)
}
\arguments{
\item{times}{Character string with \code{tabr} timing information for one or more measures. The separation of measures is done with the '|' character. Note that this is not recognized by \code{tabr}. Use the \code{times} element of the result of this function as timing information in \code{tabr}}

\item{steps}{Integer indicating the unit that the total time is reported in. E.g. 4 when the time is measured in quarter notes.}
}
\value{
Named list with in the first element \code{times} the time information without the '|' character and in the second element \code{counts} the number of units in each measure.
}
\description{
For a character string with the timing information for one or more measures the total time is calculated in units of \code{steps}. See \strong{Details} for the timing information that is recognized.
}
\section{Details}{

The following information is recognized:
\itemize{
\item an integer number like 1, 2, 4, 8, 16, 32
\item such a number with one or two dots
\item such a number with a replication factor like '*2', '*3' etc
}
}

\examples{
\dontrun{
x= check_times(" 2. | 8*3 16*4 8|8*6| 8*3 16*4 8",8)
identical(x$times,"2. 8*3 16*4 8 8*6 8*3 16*4 8")
# [1] TRUE
identical(x$counts,rep(6,4))
# [1] TRUE
}
}
