% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_transpose.R
\name{cond_transpose}
\alias{cond_transpose}
\title{transpose above with octave when below certain note}
\usage{
cond_transpose(mymusic, lowest)
}
\arguments{
\item{mymusic}{piece of music}

\item{lowest}{lowest note that will not be transposed}
}
\value{
transposed piece of music
}
\description{
Useful for e.g. guitar where e (or d) is the lowest note
}
