% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabraux.R
\name{edit_phrase}
\alias{edit_phrase}
\title{Edit a phrase}
\usage{
edit_phrase(obj, pattern, replacement, all = T)
}
\arguments{
\item{obj}{Object to change. Only objects of class \code{phrase} or \code{track} are supported}

\item{pattern}{Character string to be used as pattern in \code{sub} and \code{gsub} (depending on parameter \code{all})}

\item{replacement}{Character string to be used as replacement in \code{sub} and \code{gsub} (depending on parameter \code{all})}

\item{all}{Boolean indicating if all patterns found should be replaced (when TRUE) or only the first (when FALSE)}
}
\value{
an object of the same type as \code{obj} with possibly (i.e. when the pattern is found) a replaced 'phrase'
}
\description{
The \code{edit_phrase} is a wrapper around the R base functions \code{sub} and \code{gsub} to change the \code{phrase} in a 'musical phrase' or a 'track'
}
\section{Details}{

See the documentation for the use of \link{sub} and \link{gsub}.
}

\examples{
\dontrun{
#include a test function
}
}
