% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabraux.R
\name{expand_notes}
\alias{expand_notes}
\title{Expand notes}
\usage{
expand_notes(ns, sh_fl = 0)
}
\arguments{
\item{ns}{Character vector with notes for tabr}

\item{sh_fl}{Integer indicating the number of sharps (when positive) or flats (when negative).
In the latter case the number of flats is \code{(-sh_fl)}. When \code{ns} contains more than one character
string the value(s) in \code{sh_fl} are recycled when necessary}
}
\value{
Character vector of the same length as \code{ns} with the expanded character strings
}
\description{
A character string with incomplete note information for \code{tabr} can be expanded to the full information. See \strong{Details} for the transformations that can be applied.
}
\section{Details}{

Two transformation are applied to the notes:
\itemize{
\item when the octave number is not specified it is taken from the last one specified
\item when a note without a sharp or flat is specified it will be sharpened or flattened
according to the number of sharps and flats indicated in \code{sh_fl}.
A note that should not be sharpened or flattened can be given a additional x:
e.g. with \code{sh_fl=1} the note 'fx3' will be noted as 'f3' and not 'f#3'
}
}

\examples{
\dontrun{
expand_notes('a3 b c g f',sh_fl = 2)
expand_notes(c('a3 b c g f','a3 b c g f','a3 b c g f'),sh_fl = c(2,3))
}
}
