% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabraux.R
\name{lilypond_version}
\alias{lilypond_version}
\title{lilypond_version}
\usage{
lilypond_version()
}
\value{
Character string with the full version number (e.g. "2.19.83")
}
\description{
Retrieve the version number of Lilypond that is in use.
}
\examples{
\dontrun{
lilypond_version()
[1] "2.19.83"
}
}
