% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabraux.R
\name{pc2}
\alias{pc2}
\title{Concatenate with extra care}
\usage{
pc2(...)
}
\arguments{
\item{...}{Character strings with e.g. notes and times for tabr}
}
\description{
Simple expansion of \code{\link[tabr:pc]{tabr::pc}}: before concatenating the parts are handled by \code{\link[stringr:str_squish]{stringr::str_squish}}
}
\examples{
\dontrun{
pc2('a3 b c g  f ','  a3 b3 c4 g f')
}
}
