% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrase2.R
\name{phrase2}
\alias{phrase2}
\alias{p2}
\title{phrase with possibility of anchor points}
\usage{
phrase2(notes, info = NULL, string = NULL, bar = NULL)

p2(notes, info = NULL, string = NULL, bar = NULL)
}
\arguments{
\item{notes}{noteworthy and note info strings. When info = NULL, it is assumed that notes refers to a music object or string formatted as such.}

\item{info}{noteworthy and note info strings. When info = NULL, it is assumed that notes refers to a music object or string formatted as such.}

\item{string}{space-delimited character string or vector (or integer vector if simple string numbers). This is an optional argument that specifies which instrument strings to play for each specific timestep. Otherwise NULL.}

\item{bar}{character or NULL (default). Terminates the phrase with a bar or bar check. See Details. Also see the LilyPond help documentation on bar notation for all the valid options.}
}
\description{
The function \code{phrase2} and its alias \code{p2} extend the function \link[tabr:phrase]{tabr::phrase} (and its alias \code{p}) with the possibility to insert an anchorpoint of the form \code{'^'} or \code{'^n'} with \code{'n'} an integer. In this way it is possible When editing/copying a musical score to insert an easily recognisable point that later can be changed in e.g. a bar end.
}
\section{Details}{

See for details the documentation of \link[tabr:phrase]{tabr::phrase}. \cr If an anchorpoint is specified in \code{notes} a corresponding \code{info} element should be present, but the value is ignored. In the examples 3 notes and 1 anchorpoint are specified in \code{notes} and therefore the \code{info} parameter must describe 4 elements.\cr \cr A PR to include this functionality directly in the package  \link[tabr:tabr]{tabr} was rejected because it did not fit in the overall setup of the package.\cr Because this function is dependent on the internals of \link[tabr:tabr]{tabr} it is sensitive to changes in that package.
}

\examples{
\dontrun{
notes <- "e3 ^ f g"
info  <- "8*4"
phrase2(notes,info,bar= ":|.")
 # <Musical phrase>
 # <e>8 ^ <f>8 <g>8 \bar ":|."

notes <- "e3 ^2 f g"
info  <- "8 1 8 8"
phrase2(notes,info,bar= ":|.")
 # <Musical phrase>
 # <e>8 ^2 <f>8 <g>8 \bar ":|."
}
}
