% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabraux-package.R
\docType{package}
\name{tabraux-package}
\alias{tabraux-package}
\title{Package tabraux}
\description{
Auxiliary functions for package tabr
}
\details{
Contains two helper functions \link{pc2} and \link{expand_notes}
}
